Moved import for retrievers down a folder
