+++ examples/vector_search_with_filters.py
@@ -1,5 +1,5 @@
 from neo4j import GraphDatabase
-from neo4j_genai import VectorRetriever
+from neo4j_genai.retrievers import VectorRetriever
 
 import random
 import string
